% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_all.R
\name{build_all}
\alias{build_all}
\title{build all data files}
\usage{
build_all(
  key = NULL,
  ready_to_deploy_only = FALSE,
  silent = FALSE,
  scramble = TRUE,
  use_seed = TRUE,
  reverse_order = FALSE
)
}
\arguments{
\item{key}{optional key file (e.g. for selection of subset)}

\item{ready_to_deploy_only}{logical, only use data sets that are marked
as ready for deployment in key file. Default is \code{FALSE}, i.e.,
use all rows irrespective of their status.}

\item{silent}{logical, suppress info messages (default is \code{FALSE})}

\item{scramble}{logical (default is \code{TRUE}!!!): should the
interaction data be scrambled (see
\code{\link{read_interaction_data}} and
\code{\link{scramble_matrix}})}

\item{use_seed}{should a seed be used for scrambling that depends on the
sum of the matrix. Default is \code{TRUE}, i.e. the scrambled
matrix will be the same each time the function is called. This
only has an effect if \code{scramble = TRUE}.}

\item{reverse_order}{logical, reverse the order of entries in the key,
i.e. read the most recent additions first, which makes debugging
easier. Default is \code{FALSE}.}
}
\value{
nothing, but write files into inst/extdata/
}
\description{
read from raw data and write csv into inst/extdata/
}
