% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{check_names}
\alias{check_names}
\title{check column names and dimensions of data sets}
\usage{
check_names(xdata)
}
\arguments{
\item{xdata}{data set}
}
\value{
a vector of length 3 where the first item indicates that column names are ok (no duplicates)
 the second checks whether column and row names (if there are any) match
 the third indicates that it's a square matrix
}
\description{
check column names and dimensions of data sets
}
