% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_data_files.R
\name{delete_data_files}
\alias{delete_data_files}
\title{remove data files from package directory}
\usage{
delete_data_files()
}
\value{
a message
}
\description{
remove data files from package directory
}
