% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_project_data.R
\name{export_project_data}
\alias{export_project_data}
\title{prepare datasets linked to a specific project}
\usage{
export_project_data(project_code = NULL, output_folder = NULL, silent = TRUE)
}
\arguments{
\item{project_code}{character, the id code for the needed project}

\item{output_folder}{character, the path where the zip file should be written.
The default is the working directory.}

\item{silent}{logical (default is \code{TRUE}), suppress the progress updates
from read_subject_data and read_interaction_data.
Set to \code{FALSE} for debugging.}
}
\value{
a zip is file is written
}
\description{
prepare datasets linked to a specific project
}
\examples{
\dontrun{
export_project_data(project_code = NULL, output_folder = "~/Desktop")
}

}
