% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_file_name.R
\name{make_file_name}
\alias{make_file_name}
\title{prepare file names from key data}
\usage{
make_file_name(
  unit_code,
  cat_global,
  cat_sub = NULL,
  cat_detail = NULL,
  data_type = NULL,
  sampling_method = NULL,
  folder = "inst/extdata/",
  file_ext = ".csv"
)
}
\arguments{
\item{unit_code}{character, the required unit code}

\item{cat_global}{the main data category (as per \code{key$cat_global})}

\item{cat_sub}{the first sub category (as per \code{key$cat_sub})}

\item{cat_detail}{the second sub category (as per \code{key$cat_detail})}

\item{data_type}{additional info (as per \code{key$data_type})}

\item{sampling_method}{additional info (as per \code{key$sampling_method})}

\item{folder}{character, default is \code{"inst/extdata/"}}

\item{file_ext}{character, default is \code{".csv"}}
}
\value{
a character (i.e. file and path name)
}
\description{
prepare file names from key data
}
