% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_subjects_interactions.R
\name{match_subjects_interactions}
\alias{match_subjects_interactions}
\title{match ids in interaction data to subject-level data}
\usage{
match_subjects_interactions(unit_code = NULL, key = NULL)
}
\arguments{
\item{unit_code}{character, the required unit code (if not supplied, all
units found in \code{key} are used)}

\item{key}{the data key (if not supplied it's read from the package data)}
}
\value{
(re)writing interaction matrices and also an invisible matrix with
        reports as to who was removed and added from which table/matrix
}
\description{
match ids in interaction data to subject-level data
}
