% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_interaction_data.R
\name{read_interaction_data}
\alias{read_interaction_data}
\title{read raw interaction data sets}
\usage{
read_interaction_data(
  set = NULL,
  write = TRUE,
  key = NULL,
  scramble = FALSE,
  use_seed = FALSE,
  output_folder = "inst/extdata/"
)
}
\arguments{
\item{set}{the key numbers (or by default read everything)}

\item{write}{logical, should the read data be written to the extdata folder?
Default is \code{TRUE}. If \code{FALSE}, nothing is written and result
of reading is returned.}

\item{key}{optional data frame with the key table}

\item{scramble}{logical (default is \code{FALSE}): should the interaction
data be scrambled via \code{\link{scramble_matrix}}}

\item{use_seed}{should a seed be used for scrambling that depends on the sum of the matrix.
Default is \code{TRUE}, i.e. the scrambled matrix will be the same
each time the function is called. This only has an effect if
\code{scramble = TRUE}.}

\item{output_folder}{character, the path where the csv files should be written.
The default is inside the package directory
(\code{"inst/extdata/"}).}
}
\value{
writes files and returns a message. Or returns an interaction matrix.
}
\description{
read raw interaction data sets
}
