% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_subject_data.R
\name{read_subject_data}
\alias{read_subject_data}
\title{read subject data (observation time, age, sex etc)}
\usage{
read_subject_data(
  set = NULL,
  write = TRUE,
  key = NULL,
  silent = FALSE,
  output_folder = "inst/extdata/"
)
}
\arguments{
\item{set}{the key numbers (or by default read every subject data found)}

\item{write}{logical, should the read data be written to the \code{extdata}
folder? Default is \code{TRUE}. If \code{FALSE}, nothing is
written and result of reading is returned as data.frame.}

\item{key}{optional data frame with the key table}

\item{silent}{logical, suppress info messages (default is \code{FALSE})}

\item{output_folder}{character, the path where the csv files should be written.
The default is inside the package directory
(\code{"inst/extdata/"}).}
}
\value{
writes files and returns a message (or a data.frame)
}
\description{
read subject data (observation time, age, sex etc)
}
