% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_accents.R
\name{remove_accents}
\alias{remove_accents}
\title{remove and replace accents in character vectors}
\usage{
remove_accents(x)
}
\arguments{
\item{x}{character vector}
}
\value{
x with accents replaced
}
\description{
remove and replace accents in character vectors
}
\examples{
remove_accents(c("Émile", "André", "Bör", "Taï"))
}
