% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scramble_matrix.R
\name{scramble_matrix}
\alias{scramble_matrix}
\title{data scrambling for pseudo-blinding}
\usage{
scramble_matrix(mat, use_seed = TRUE)
}
\arguments{
\item{mat}{square matrix to be scrambled}

\item{use_seed}{should a seed be used that depends on the sum of the matrix.
Default is \code{TRUE}, i.e. the scrambled matrix will be the same
each time the function is called.}
}
\value{
scrambled input matrix
}
\description{
data scrambling for pseudo-blinding
}
\details{
The scrambling works by first selecting a random dyad for which at
         least one its two values is not zero. It then selects one more dyad,
         which may or may not be a zero-dyad. It then assigns the two
         values of the first dyad to the cells of the second dyad,
         and vice versa.
         As such, the total sum of a matrix is kept constant, but
         row and column sums will change from the input matrix.
}
