% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_keyfile.R
\name{test_keyfile}
\alias{test_keyfile}
\title{test internal consistency of the key file}
\usage{
test_keyfile()
}
\value{
textual messages (and invisible logical vector)
}
\description{
test internal consistency of the key file
}
